# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentation pour MyHAP API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "njgconnect@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/dossier/{idDossier}": {
      "get": {
        "tags": [
          "Dossier"
        ],
        "summary": "Informations d'un Dossier en fonction de son id",
        "description": "Renvoi un Object contenant les infos MyHAP & Analysimmo & (en option les Techniciens)",
        "operationId": "getDossierById",
        "parameters": [
          {
            "name": "idDossier",
            "in": "path",
            "description": "ID du dossier recherché",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "employesParam",
            "in": "query",
            "description": "Affiche tout les Techniciens en réponse ( avec le name & id )",
            "required": false,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opération Réussi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dossier"
                }
              }
            }
          },
          "400": {
            "description": "Une erreur est survenue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Veuillez renseigner un id pour trouver votre dossier",
                    "Aucun dossier n'a été renseigné avec cette id",
                    "Erreur, (nom de l'erreur)"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Dossier"
        ],
        "summary": "Modifie les Informations d'un Dossier en fonction de son id",
        "description": "Permet la modification d'un dossier  MyHAP & Analysimmo.",
        "operationId": "updateDossierById",
        "parameters": [
          {
            "name": "idDossier",
            "in": "path",
            "description": "ID du dossier recherché",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Information nécéssaire",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/DossierUpdateField"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Opération Réussi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dossier"
                }
              }
            }
          },
          "400": {
            "description": "Une erreur est survenue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Veuillez renseigner un id pour trouver votre dossier",
                    "Veuillez renseigner un id valable pour trouver votre dossier",
                    "Aucun dossier n'a été renseigné avec cette id",
                    "Erreur, (nom de l'erreur)"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/dossier/{idDossier}/intervention/{idIntervention}/prelevement/{idPrelevement}": {
      "patch": {
        "tags": [
          "Prelevement"
        ],
        "summary": "Modifie le Prélevement & couche lié ",
        "description": "Permet la modification d'un Prélevement accompagné de la Modification | Suppréssion | Création de couche",
        "operationId": "updatePrelevementById",
        "parameters": [
          {
            "name": "idDossier",
            "in": "path",
            "description": "ID du dossier lié au prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idIntervention",
            "in": "path",
            "description": "ID de l'intervention raccroché au prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idPrelevement",
            "in": "path",
            "description": "ID du prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Pour le champ **'couches'** Si l'object **ne contient que l'id** alors la couche sera **Delete** | Si il y a des données mais **pas d'id** alors la couche sera **créer** | si il y a **l'id et les infos** alors c'est une **modification**",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/PrelevementField"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Opération Réussi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dossier"
                }
              }
            }
          },
          "400": {
            "description": "Une erreur est survenue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Veuillez renseigner un id Dossier | Intervention | prelevement pour trouver votre intervention",
                    "Veuillez renseigner un id Dossier | Intervention | prelevement valable pour trouver votre intervention",
                    "Veuillez renseigner tout les champs pour modifier votre Prélèvement",
                    "Votre Intervention n'existe pas",
                    "Votre Prelevement n'existe pas",
                    "Une erreur est survenue lors de l'édition de votre Prélèvement"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Prelevement"
        ],
        "summary": "Supprime le Prélevement & les couches liées ",
        "description": "Permet la Suppréssion d'un Prélevement accompagné de ses couches",
        "operationId": "deletePrelevementById",
        "parameters": [
          {
            "name": "idDossier",
            "in": "path",
            "description": "ID du dossier lié au prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idIntervention",
            "in": "path",
            "description": "ID de l'intervention raccroché au prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idPrelevement",
            "in": "path",
            "description": "ID du prélevement",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opération Réussi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dossier"
                }
              }
            }
          },
          "400": {
            "description": "Une erreur est survenue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Veuillez renseigner un id Dossier | Intervention | prelevement pour trouver votre intervention",
                    "Veuillez renseigner un id Dossier | Intervention | prelevement valable pour trouver votre intervention",
                    "Veuillez renseigner tout les champs pour modifier votre Prélèvement",
                    "Votre Prelevement n'existe pas",
                    "Problème dans la suppression des couches lié à votre prélèvement",
                    "Problème dans la suppression du Prelevement."
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "FileTypeEnum": {
        "type": "string",
        "enum": [
          "ANNEXE",
          "PRE_RAPPORT",
          "RAPPORT",
          "PLAN",
          "DICT",
          "INTERVENTION",
          "DOSSIER"
        ]
      },
      "Dossier": {
        "type": "object",
        "properties": {
          "diag": {
            "type": "object",
            "properties": {
              "numero": {
                "type": "number"
              },
              "idDossier": {
                "type": "number"
              },
              "reference": {
                "type": "string"
              },
              "adresse": {
                "type": "string"
              },
              "cptAdresse": {
                "type": "string"
              },
              "ville": {
                "type": "string"
              },
              "pays": {
                "type": "string"
              },
              "departement": {
                "type": "string"
              },
              "longitude": {
                "type": "string"
              },
              "latitude": {
                "type": "string"
              },
              "commentaire": {
                "type": "string"
              },
              "dateCommande": {
                "type": "string",
                "format": "date-time",
                "example": "2018-10-19T09:52:58.336Z"
              },
              "StatutDossier": {
                "type": "object",
                "properties": {
                  "intitule": {
                    "type": "string",
                    "example": "Terminé",
                    "enum": [
                      "A contacter",
                      "A planifier",
                      "En attente d'expertise",
                      "En attente de résultats",
                      "A facturer",
                      "En attente BC / BT / OS",
                      "Demande incomplète",
                      "Rapport à envoyer",
                      "En attente rédaction rapport",
                      "En attente de règlement",
                      "En attente Visa 1",
                      "Bannette PROBLEME",
                      "En attente résultats HAP",
                      "En attente de ré-intervention",
                      "Annulé",
                      "Terminé",
                      "En attente de Rapport Final",
                      "Données non exportées",
                      "En attente facturation",
                      "En attente Visa 2",
                      "Sous-traités",
                      "En attente ré-intervention BCL",
                      "Litige",
                      "En attente dépôt prélèvements",
                      "StandBy Bailleurs CC",
                      "Stand-By 1001VH",
                      "En attente infos/documents - DPE",
                      "DRIVE"
                    ]
                  }
                }
              },
              "idStatut": {
                "type": "number"
              },
              "idEmployeIntervention": {
                "type": "number"
              },
              "codePostal": {
                "type": "string"
              },
              "docs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "link": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "$ref": "#/components/schemas/FileTypeEnum"
                    }
                  }
                }
              }
            }
          },
          "myHAP": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "isParkMarker": {
                "type": "boolean"
              },
              "typologie": {
                "type": "string"
              },
              "docs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "link": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "date": {
                      "type": "date-time"
                    },
                    "type": {
                      "$ref": "#/components/schemas/FileTypeEnum"
                    }
                  }
                }
              },
              "interventions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Intervention"
                }
              }
            }
          }
        }
      },
      "Intervention": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "dateDebutMission": {
            "type": "string",
            "format": "date-time"
          },
          "dateFinMission": {
            "type": "string",
            "format": "date-time"
          },
          "idEmployeIntervention": {
            "type": "number"
          },
          "zones": {
            "type": "string"
          },
          "isFirstIntervention": {
            "type": "boolean"
          },
          "idDossier": {
            "type": "number"
          },
          "prelevements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prelevement"
            }
          }
        }
      },
      "Prelevement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "numero": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "emplacement": {
            "type": "string"
          },
          "codePostal": {
            "type": "number"
          },
          "ville": {
            "type": "string"
          },
          "departement": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "float"
          },
          "longitude": {
            "type": "number",
            "format": "float"
          },
          "altitude": {
            "type": "number",
            "format": "float"
          },
          "adresse": {
            "type": "string"
          },
          "materiaux": {
            "type": "string"
          },
          "taille": {
            "type": "number"
          },
          "couleur": {
            "type": "string"
          },
          "images": {
            "type": "string"
          },
          "laboratoire": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "resultat": {
            "type": "string"
          },
          "PrelevementPossible": {
            "type": "boolean"
          },
          "choixPrelevementImPossible": {
            "type": "string"
          },
          "idIntervention": {
            "type": "number"
          },
          "couches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Couche"
            }
          }
        }
      },
      "Couche": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "numero": {
            "type": "number"
          },
          "taille": {
            "type": "number"
          },
          "couleur": {
            "type": "string"
          },
          "amiante": {
            "type": "boolean"
          },
          "materiaux": {
            "type": "string"
          },
          "HAP": {
            "type": "string"
          },
          "liant": {
            "type": "boolean"
          },
          "granulat": {
            "type": "boolean"
          },
          "idPrelevement": {
            "type": "number"
          },
          "laboratoire": {
            "type": "string"
          },
          "bonCommandeLabo": {
            "type": "string"
          }
        }
      }
    },
    "requestBodies": {
...
      "DossierUpdateField": {
        "type": "object",
        "properties": {
          "diag": {
            "type": "object",
            "properties": {
              "reference": {
                "type": "string"
              },
              "adresse": {
                "type": "string"
              },
              "cptAdresse": {
                "type": "string"
              },
              "ville": {
                "type": "string"
              },
              "pays": {
                "type": "string"
              },
              "departement": {
                "type": "string"
              },
              "longitude": {
                "type": "string"
              },
              "latitude": {
                "type": "string"
              },
              "commentaire": {
                "type": "string"
              },
              "dateCommande": {
                "type": "string",
                "format": "date-time",
                "example": "2018-10-19T09:52:58.336Z"
              },
              "idStatut": {
                "type": "number"
              }
            }
          },
          "myHAP": {
            "type": "object",
            "properties": {
              "isParkMarker": {
                "type": "boolean"
              },
              "typologie": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PrelevementField": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "emplacement": {
            "type": "string"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "altitude": {
            "type": "string"
          },
          "adresse": {
            "type": "string"
          },
          "materiaux": {
            "type": "string"
          },
          "taille": {
            "type": "number"
          },
          "couleur": {
            "type": "string"
          },
          "images": {
            "type": "string"
          },
          "laboratoire": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "resultat": {
            "type": "string"
          },
          "PrelevementPossible": {
            "type": "boolean"
          },
          "choixPrelevementImPossible": {
            "type": "string"
          },
          "couches": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "taille": {
                  "type": "string"
                },
                "couleur": {
                  "type": "string"
                },
                "amiante": {
                  "type": "boolean"
                },
                "materiaux": {
                  "type": "string"
                },
                "HAP": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}